requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0
numpy>=1.24.0
Pillow>=10.0.0
# TensorFlow is optional now (for backward compatibility)
# tensorflow>=2.13.0

